fix: Norminette ok for the project
