+++ execution/ft_input.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_input.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mwiecek <mwiecek@student.42.fr>            +#+  +:+       +#+        */
+/*   By: zslowian <zslowian@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/28 18:48:04 by zslowian          #+#    #+#             */
-/*   Updated: 2025/05/05 18:55:32 by mwiecek          ###   ########.fr       */
+/*   Updated: 2025/05/05 19:00:15 by zslowian         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -65,43 +65,6 @@ void	ft_execute_cmd(t_global *g, t_command *cmd, pid_t prev_pid)
 			waitpid(cmd->cmd_pid, &wstatus, 0);
 			if (WIFEXITED(wstatus))
 				g->last_exit_code = WEXITSTATUS(wstatus);
-		}
-	}
-}
-
-void	ft_check_error(char *path, int *error, bool *dir_or_cmd)
-{
-	if (*error && *dir_or_cmd)
-		return (ft_handle_minishell_err(path, ": No such file or directory\n"));
-	if (*error)
-	{
-		if (*dir_or_cmd)
-		{
-			*error = 126;
-			return (ft_handle_minishell_err(path,
-					": No such file or directory\n"));
-		}
-		else
-			return (ft_handle_minishell_err(path, ": command not found\n"));
-	}
-}
-
-void	ft_check_error2(char* path, int *error, bool *dir_or_cmd)
-{
-	if (access(path, F_OK))
-	{
-		*error = 127;
-		ft_minishell_perror(path, ENOENT);
-	}
-	else if (access(path, X_OK))
-	{
-		*error = 126;
-		if (dir_or_cmd)
-			ft_minishell_perror(path, EACCES);
-		else
-		{
-			*error = 127;
-			return (ft_handle_minishell_err(path, ": command not found\n"));
 		}
 	}
 }
